feat(vlayer update): insert version to prover URL.
