+++ rust/cli/src/utils/mod.rs
@@ -1,2 +1,3 @@
 pub(crate) mod parse_toml;
 pub(crate) mod path;
+pub(crate) mod url;
